+++ src/Pwa/ShowInstallConfirm.mjs
@@ -19,9 +19,9 @@ export class ShowInstallConfirm {
     /**
      * @param {Localization} localization
      * @param {StyleSheetManager | null} style_sheet_manager
-     * @returns {ShowInstallConfirm}
+     * @returns {Promise<ShowInstallConfirm>}
      */
-    static new(localization, style_sheet_manager = null) {
+    static async new(localization, style_sheet_manager = null) {
         return new this(
             localization,
             style_sheet_manager